+++ tests/testthat/test_aorsf_classif_aorsf.R
@@ -1,10 +1,8 @@
 test_that("autotest", {
-  learner = lrn("classif.aorsf")
-  expect_learner(learner)
-  result = run_autotest(
-    learner,
-    exclude = "utf8_feature_names",
-    N = 30
-  )
-  expect_true(result, info = result$error)
+  with_seed(42, {
+    learner = lrn("classif.aorsf", n_tree = 20, na_action = "impute_meanmode")
+    expect_learner(learner)
+    result = run_autotest(learner, N = 30, exclude = "utf8_feature_names")
+    expect_true(result)
+  })
 })