+++ tests/testthat/test_paramtest_aorsf_classif_aorsf.R
@@ -31,9 +31,8 @@ test_that("paramtest classif.aorsf train", {
 
 test_that("paramtest classif.aorsf predict", {
   learner = lrn("classif.aorsf")
-  fun_list = list(aorsf:::predict.ObliqueForest)
+  fun_list = list(aorsf:::predict.ObliqueForest) # nolint
   exclude = c(
-    "na_action", # not implemented,
     "new_data", # handled internally
     "object", # handled internally
     "pred_horizon", # all times required for distr